FASTAPISERVER/requirements.txt
# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Pydantic for data validation
pydantic==2.10.6
pydantic-settings==2.7.1

# LangChain and AI
langchain==0.3.14
langchain-google-genai==2.0.8
langgraph==0.2.64

# Google Generative AI
google-generativeai==0.8.3

# Image processing
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==11.0.0

# Web scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Environment variables
python-dotenv==1.0.1

# Rich console output (optional, for debugging)
rich==13.9.4
